Refactor the client parser and try to detect the `directive`
In client `parseDOM()`, update the regex match of the html string
to also detect the directive or doctype. If doctype is found,
pass the string as the directive parameter in `formatDOM()`.

Improve dom parsing logic by creating a document for top-level
nodes like `html`, `head`, and `body`. In all other cases, create
a `body` element container and use `innerHTML`.