fix(lib): improve head and body regex in `domparser.js`
Because the head and body regexes test against the closing tag,
this causes html with unclosed head or body to not be parsed
correctly.

For example, given the following:

```js
parse('<html><body>');
```

The expected output is:

```
[ { type: 'tag',
    name: 'html',
    attribs: {},
    children:
     [ { type: 'tag',
         name: 'body',
         attribs: {},
         children: [],
         next: null,
         prev: null,
         parent: [Circular] } ],
    next: null,
    prev: null,
    parent: null } ]
```

But the actual output is:

```
[
  {
    "next": null,
    "prev": null,
    "parent": null,
    "name": "html",
    "attribs": {},
    "type": "tag",
    "children": []
  }
]
```

The fix is to update the regex to use the opening tag instead of
the closing tag.

Add test case.

Fixes #18